+++ CHANGELOG.md
@@ -10,3 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update to Minecraft 26.2.x
+
+### Fixed
+
+- Fix unable to pick parrot spawn egg from the custom parrot mob